app/telegram/bot/requirements.txt
python_telegram_bot==20.0
telegram==0.0.1
python-dotenv==0.21.0